refactor(skeleton): 优化目录结构
