+++ packages/devui-vue/devui/skeleton/index.ts
@@ -1,10 +1,10 @@
 import type { App } from 'vue';
 import Skeleton from './src/skeleton';
-import SkeletonItem from './src/item/item';
+import SkeletonItem from './src/components/skeleton-item';
 
 export * from './src/skeleton-types';
 
-export { Skeleton,SkeletonItem };
+export { Skeleton, SkeletonItem };
 
 export default {
   title: 'Skeleton 骨架屏',
@@ -12,6 +12,6 @@ export default {
   status: '100%',
   install(app: App): void {
     app.component(Skeleton.name, Skeleton);
-  app.component(SkeletonItem.name, SkeletonItem);
-  }
+    app.component(SkeletonItem.name, SkeletonItem);
+  },
 };
