+++ packages/devui-vue/devui/skeleton/src/components/skeleton-item-types.ts
@@ -5,35 +5,35 @@ export type ModelValue = number | string;
 export const itemProps = {
   row: {
     type: Number,
-    default: 0
+    default: 0,
   },
   animate: {
     type: Boolean,
-    default: true
+    default: true,
   },
   round: {
     type: Boolean,
-    default: false
+    default: false,
   },
   loading: {
     type: Boolean,
-    default: true
+    default: true,
   },
   avatarShape: {
     type: String as PropType<'round' | 'square'>,
-    default: 'round'
+    default: 'round',
   },
   titleWidth: {
     type: [String, Number] as PropType<ModelValue>,
-    default: '40%'
+    default: '40%',
   },
   rowWidth: {
     type: [Number, String, Array] as PropType<number | string | Array<number | string>>,
-    default: ['100%']
+    default: ['100%'],
   },
   shape: {
     type: String as PropType<'avatar' | 'image' | 'title' | 'paragraph' | 'button'>,
-  }
+  },
 } as const;
 
 export type ItemProps = ExtractPropTypes<typeof itemProps>;
