+++ packages/devui-vue/devui/skeleton/src/components/skeleton-item.scss
