+++ packages/devui-vue/devui/skeleton/src/skeleton-types.ts
@@ -5,48 +5,48 @@ export type ModelValue = number | string;
 export const skeletonProps = {
   row: {
     type: Number,
-    default: 0
+    default: 0,
   },
   animate: {
     type: Boolean,
-    default: true
+    default: true,
   },
   round: {
     type: Boolean,
-    default: false
+    default: false,
   },
   loading: {
     type: Boolean,
-    default: true
+    default: true,
   },
   avatar: {
     type: Boolean,
-    default: false
+    default: false,
   },
   title: {
     type: Boolean,
-    default: true
+    default: true,
   },
   paragraph: {
     type: Boolean,
-    default: true
+    default: true,
   },
   avatarSize: {
     type: [String, Number] as PropType<ModelValue>,
-    default: '40px'
+    default: '40px',
   },
   avatarShape: {
     type: String as PropType<'round' | 'square'>,
-    default: 'round'
+    default: 'round',
   },
   titleWidth: {
     type: [String, Number] as PropType<ModelValue>,
-    default: '40%'
+    default: '40%',
   },
   rowWidth: {
     type: [Number, String, Array] as PropType<number | string | Array<number | string>>,
-    default: ['100%']
-  }
+    default: ['100%'],
+  },
 } as const;
 
 export type SkeletonProps = ExtractPropTypes<typeof skeletonProps>;
