+++ packages/devui-vue/devui/skeleton/src/skeleton.tsx
@@ -1,7 +1,6 @@
-import './skeleton.scss';
-
 import { defineComponent } from 'vue';
 import { skeletonProps, SkeletonProps } from './skeleton-types';
+import './skeleton.scss';
 
 export default defineComponent({
   name: 'DSkeleton',
@@ -28,11 +27,11 @@ export default defineComponent({
           for (let index = 0; index < rowNum; index++) {
             if (rowWidth[index]) {
               switch (typeof rowWidth[index]) {
-              case 'string':
-                arr.push({ width: rowWidth[index] });
-                break;
-              case 'number':
-                arr.push({ width: `${rowWidth[index]}px` });
+                case 'string':
+                  arr.push({ width: rowWidth[index] });
+                  break;
+                case 'number':
+                  arr.push({ width: `${rowWidth[index]}px` });
               }
             } else {
               arr.push({ width: 1 });
@@ -40,21 +39,23 @@ export default defineComponent({
           }
         } else {
           switch (typeof rowWidth) {
-          case 'string':
-            pushIntoArray(rowWidth);
-            break;
-          case 'number':
-            pushIntoArray(`${rowWidth}px`);
-            break;
+            case 'string':
+              pushIntoArray(rowWidth);
+              break;
+            case 'number':
+              pushIntoArray(`${rowWidth}px`);
+              break;
           }
         }
       })();
 
-      return <div class="devui-skeleton__paragraph" v-show={isShown}>{
-        arr.map(item => {
-          return <div class="devui-skeleton__item" style={round ? 'border-radius: 1em;' : '' + `width: ${item.width}`} />;
-        })
-      }</div>;
+      return (
+        <div class="devui-skeleton__paragraph" v-show={isShown}>
+          {arr.map((item) => {
+            return <div class="devui-skeleton__item" style={round ? 'border-radius: 1em;' : '' + `width: ${item.width}`} />;
+          })}
+        </div>
+      );
     }
 
     function renderAvatarStyle(avatarSize, avatarShape) {
@@ -63,49 +64,49 @@ export default defineComponent({
       }
       function renderAvatarSize() {
         switch (typeof avatarSize) {
-        case 'string':
-          return `width:${avatarSize};height:${avatarSize};`;
-        case 'number':
-          return `width:${avatarSize}px;height:${avatarSize}px;`;
+          case 'string':
+            return `width:${avatarSize};height:${avatarSize};`;
+          case 'number':
+            return `width:${avatarSize}px;height:${avatarSize}px;`;
         }
       }
 
-      return (renderAvatarSize(avatarSize) + renderAvatarShape(avatarShape));
+      return renderAvatarSize(avatarSize) + renderAvatarShape(avatarShape);
     }
     function renderTitle(isVisible, titleWidth, isRound) {
       function renderTitleWidth() {
         switch (typeof titleWidth) {
-        case 'string':
-          return `width: ${titleWidth};`;
-        case 'number':
-          return `width: ${titleWidth}px;`;
+          case 'string':
+            return `width: ${titleWidth};`;
+          case 'number':
+            return `width: ${titleWidth}px;`;
         }
       }
       function renderTitleVisibility() {
         return isVisible ? null : 'visibility: hidden;';
       }
 
-      return (renderTitleWidth(titleWidth) + renderBorderRadius(isRound) + renderTitleVisibility(isVisible));
+      return renderTitleWidth(titleWidth) + renderBorderRadius(isRound) + renderTitleVisibility(isVisible);
     }
     function renderDefaultSkeleton() {
-      return <>
-        <div class="devui-skeleton__avatar" v-show={props.avatar}>
-          <div class="avatar" style={renderAvatarStyle(props.avatarSize, props.avatarShape)} />
-        </div>
-        <div class="devui-skeleton__group">
-          <div class="devui-skeleton__title" style={renderTitle(props.title, props.titleWidth, props.round)} />
-          {renderParagraph(props.paragraph, props.row, props.rowWidth, props.round)}
-        </div>
-      </>;
+      return (
+        <>
+          <div class="devui-skeleton__avatar" v-show={props.avatar}>
+            <div class="avatar" style={renderAvatarStyle(props.avatarSize, props.avatarShape)} />
+          </div>
+          <div class="devui-skeleton__group">
+            <div class="devui-skeleton__title" style={renderTitle(props.title, props.titleWidth, props.round)} />
+            {renderParagraph(props.paragraph, props.row, props.rowWidth, props.round)}
+          </div>
+        </>
+      );
     }
 
     return () => {
       if (props.loading) {
-        return <div class={`devui-skeleton ${renderAnimate(props.animate)}`}>
-          {renderDefaultSkeleton()}
-        </div>;
+        return <div class={`devui-skeleton ${renderAnimate(props.animate)}`}>{renderDefaultSkeleton()}</div>;
       }
       return <>{slots.default?.()}</>;
     };
-  }
+  },
 });
